Moved ProductOverview files around, used John's Star for average ratings, and changed ProductOverview padding and margin. Still has CSS borders for debugging purposes
