+++ client/src/ProductOverview/ProductInformation.jsx
@@ -1,15 +1,17 @@
-import React from 'react';
-import StarRating from './StarRating.jsx';
+import React, {useContext} from 'react';
+import { Stars } from '../helper/Stars.jsx';
+import { MetaContext } from '../context.js';
 
 export default function ProductInformation({style, product}) {
   const {category, name, default_price, slogan} = product;
   const {original_price, sale_price} = style;
 
-
+  const ratings = useContext(MetaContext);
 
   return (
     <div>
-      <StarRating />
+      <Stars rating={ratings}/>
+      <a href="" onClick={(e) => e.preventDefault()}>Read All Reviews</a>
       <div>{category}</div>
       <div>{name}</div>
       {sale_price ?
